+++ packages/typera-koa/tests/routeHandler.spec.ts
@@ -81,7 +81,7 @@ describe('routeHandler', () => {
       _: Koa.Context
     ) =>
       new Promise((resolve, _reject) => {
-        setTimeout(() => resolve(Either.right({ foo: 42 })), 10)
+        setTimeout(() => resolve(Middleware.next({ foo: 42 })), 10)
       })
 
     const handler: RouteHandler<